+++ src/deploy/demodata/V1000__demo_data.sql
@@ -1,0 +1,3 @@
+INSERT INTO country (id, isocode, name, information, secret) values('1', 'de', 'Germany', 'North Europe', null);
+INSERT INTO country (id, isocode, name, information, secret) values('2', 'es', 'Spain', 'West Europe', null);
+INSERT INTO country (id, isocode, name, information, secret) values('3', 'it', 'Italy', 'South Europe', null);