+++ src/main/java/org/goafabric/spring/boot/exampleservice/persistence/provisioning/DatabaseImport.java
@@ -5,5 +5,4 @@ package org.goafabric.spring.boot.exampleservice.persistence.provisioning;
  */
 interface DatabaseImport {
     void importCatalogData(String... args);
-    void importDemoData(String... args);
 }