+++ src/main/resources/db/migration/V1__init.sql
@@ -5,14 +5,9 @@ create table country
 	id varchar(36) not null
 		constraint pk_country
 			primary key,
-    tenantid varchar(3) not null,
 	isocode varchar(2),
 	name varchar(255),
 	information varchar(255),
 	secret varchar(2048), -- The field has to be much bigger because it has to hold encrypted info
 	version bigint default 0
 );
-
-INSERT INTO country (id, tenantid, isocode, name, information, secret) values('1', '10', 'de', 'Germany', 'North Europe', null);
-INSERT INTO country (id, tenantid, isocode, name, information, secret) values('2', '10','es', 'Spain', 'West Europe', null);
-INSERT INTO country (id, tenantid,isocode, name, information, secret) values('3', '10','it', 'Italy', 'South Europe', null);