+++ doc/release-notes.md
@@ -6,10 +6,15 @@ CrossFunctional
 - Postgres Database Monitoring
     - @see InfoActuator
 - Fixed Web endpoints inside static\index.html not beeing relative
+-
 
 Security
 - Encrypted Searchable Fields in Database (less secure)
     - @See: EncryptionConfiguration + @encryptedSearchableString
+
+Persistence
+- DemoData added to DatabaseProvisioning
+
     
 Docker
 - Nginx Reverse Proxy for HTTP and HTTPS